Harmonize deployment logic
Consolidate deployment logic and types.  Remove duplicated code.
Improve types, including their names.  Move all configuration values
under `sdm.k8`.

For side-effect goal fulfillments, discern on current package name
rather than hard-coded "@atomist/k8-automation".

Use Commit.images in GraphQL rather than deprecated Commit.image.

Reorganize and rename files to reflect their purpose.  Move Kubernetes
types into their own file.

Migrate to async style code from promise chains.

Update TSLint configuration and fix issues.

Closes #16
