+++ lib/support/deploy.ts
@@ -15,13 +15,21 @@
  */
 
 import {
+    HandlerResult,
+    logger,
+} from "@atomist/automation-client";
+import {
     ExecuteGoal,
     ExecuteGoalResult,
     GoalInvocation,
+    ProgressLog,
     SdmGoalEvent,
 } from "@atomist/sdm";
 import stringify = require("json-stringify-safe");
 import * as k8 from "kubernetes-client";
+import {
+    defaultNamespace,
+} from "../typings/kubernetes";
 import {
     endpointBaseUrl,
     getKubeConfig,
@@ -29,6 +37,9 @@ import {
     KubeApplicationRequest,
     upsertApplication,
 } from "./api";
+import {
+    KubernetesGoalData,
+} from "./goal";
 
 export interface KubernetesDeployerOptions {
     context: string;
@@ -38,7 +49,7 @@ export function executeKubernetesDeploy(): ExecuteGoal {
     return async (goalInvocation: GoalInvocation): Promise<ExecuteGoalResult> => {
         const { configuration, context, sdmGoal, progressLog } = goalInvocation;
 
-        const options = configuration.sdm.k8 as KubernetesDeployerOptions;
+        const options: KubernetesDeployerOptions = configuration.sdm.k8;
         const repo = sdmGoal.repo.name;
         const owner = sdmGoal.repo.owner;
         const sha = sdmGoal.sha;
@@ -48,7 +59,7 @@ export function executeKubernetesDeploy(): ExecuteGoal {
         if (!sdmGoal.push.after.image) {
             const msg = `Kubernetes deploy requested for ${depName} but that commit ` +
                 `has no Docker image associated with it`;
-            return { code: 1, message: msg };
+            return failDeploy(progressLog, msg);
         }
         const image = sdmGoal.push.after.image.imageName;
         progressLog.write(`Processing ${depName}`);
@@ -57,71 +68,81 @@ export function executeKubernetesDeploy(): ExecuteGoal {
         try {
             k8Config = getKubeConfig(options.context);
         } catch (e) {
-            return { code: 1, message: e.message };
+            return failDeploy(progressLog, `Failed to get kube config for ${depName}: ${e.message}`);
         }
 
-        let kubeApp: KubeApplication;
-        try {
-            kubeApp = validateSdmGoal(sdmGoal);
-        } catch (e) {
-            const msg = `${depName} ${e.message}`;
-            return { code: 1, message: msg };
+        const kubeGoalData = validateSdmGoal(sdmGoal);
+        if (!kubeGoalData) {
+            return { code: 1, message: "No Kubernetes goal data found" };
         }
-        if (!kubeApp) {
-            return { code: 1, message: "No Kubernetes deployment data found" };
-        }
+
+        const kubeApp: KubeApplication = {
+            ...kubeGoalData,
+            workspaceId,
+            image,
+            ns: kubeGoalData.ns || defaultNamespace,
+        };
 
         progressLog.write(`Deploying ${depName} to Kubernetes`);
         const upsertReq: KubeApplicationRequest = {
             ...kubeApp,
             config: k8Config,
-            workspaceId,
-            image,
         };
-        return upsertApplication(upsertReq)
-            .then(() => {
-                const message = `Successfully deployed ${depName} to Kubernetes`;
-                const description = `Deployed to Kubernetes namespace \`${kubeApp.ns}\``;
-                progressLog.write(message);
-                if (kubeApp.path && kubeApp.host) {
-                    return { code: 0, message, targetUrl: endpointBaseUrl(kubeApp), description };
-                } else {
-                    return { code: 0, message, description };
-                }
-            }, e => {
-                const msg = `Failed to deploy ${depName} to Kubernetes: ${e.message}`;
-                progressLog.write(msg);
-                return { code: 1, message: msg };
-            });
+        try {
+            await upsertApplication(upsertReq);
+        } catch (e) {
+            return failDeploy(progressLog, `Failed to deploy ${depName} to Kubernetes: ${e.message}`);
+        }
+        const message = `Successfully deployed ${depName} to Kubernetes`;
+        progressLog.write(message);
+        const description = `Deployed to Kubernetes namespace \`${kubeApp.ns}\``;
+        const label = `Kubernetes ${env}:${kubeApp.ns}`;
+        const targetUrls = (kubeApp.path && kubeApp.host) ? [{ label, url: endpointBaseUrl(kubeApp) }] : undefined;
+        return { code: 0, message, description, targetUrls };
     };
 }
 
 /**
- * Validate the SDM goal has all necessary data.  It will throw an
- * Error if the goal is invalid in some way.  It will return undefined
- * if nothing should be deployed.
+ * Validate the SDM goal has all necessary data.  It will return the
+ * KubeApplication data if the deployment should proceed.  It will ll
+ * return undefined if nothing should be deployed.
  *
  * @param sdmGoal SDM goal for Kubernetes application deployment
  * @return valid KubeApplication if something should be deployed,
  *         undefined if nothing should be deployed
  */
-export function validateSdmGoal(sdmGoal: SdmGoalEvent): KubeApplication {
+export function validateSdmGoal(sdmGoal: SdmGoalEvent): KubernetesGoalData | undefined {
     if (!sdmGoal.data) {
-        throw new Error(`SDM goal data property is false, cannot deploy: '${stringify(sdmGoal)}'`);
+        logger.debug(`SDM goal data property is false, cannot deploy: '${stringify(sdmGoal)}'`);
+        return undefined;
     }
     let sdmData: any;
     try {
         sdmData = JSON.parse(sdmGoal.data);
     } catch (e) {
-        e.message = `Failed to parse SDM goal data '${sdmGoal.data}' as JSON: ${e.message}`;
-        throw e;
+        logger.debug(`Failed to parse SDM goal data '${sdmGoal.data}' as JSON: ${e.message}`);
+        return undefined;
     }
     if (!sdmData.kubernetes) {
-        throw new Error(`SDM goal data kubernetes property is false, cannot deploy: '${stringify(sdmData)}'`);
+        logger.debug(`SDM goal data kubernetes property is false, cannot deploy: '${stringify(sdmData)}'`);
+        return undefined;
     }
-    const kubeApp: KubeApplication = sdmData.kubernetes;
+    const kubeApp: KubernetesGoalData = sdmData.kubernetes;
     if (!kubeApp.name) {
-        throw new Error(`SDM goal data kubernetes name property is false, cannot deploy: '${stringify(sdmData)}'`);
+        logger.debug(`SDM goal data kubernetes name property is false, cannot deploy: '${stringify(sdmData)}'`);
+        return undefined;
     }
     return kubeApp;
+}
+
+/**
+ * Report and return failure.
+ *
+ * @param message informative error message
+ * @return a failure handler result using the provided error message
+ */
+function failDeploy(log: ProgressLog, message: string): HandlerResult {
+    log.write(message);
+    logger.error(message);
+    return { code: 1, message };
 }
