+++ package.json
@@ -27,11 +27,17 @@
   "main": "./index.js",
   "types": "./index.d.ts",
   "dependencies": {
+    "@types/app-root-path": "^1.2.4",
+    "@types/fs-extra": "^5.0.4",
+    "@types/retry": "^0.10.2",
+    "app-root-path": "^2.1.0",
     "docker-file-parser": "^1.0.3",
+    "fs-extra": "^7.0.1",
     "json-stringify-safe": "^5.0.1",
     "kubernetes-client": "^5.3.1",
     "lodash": "^4.17.11",
-    "promise-retry": "^1.1.1"
+    "promise-retry": "^1.1.1",
+    "retry": "^0.12.0"
   },
   "peerDependencies": {
     "@atomist/automation-client": "*",
@@ -49,15 +55,13 @@
     "espower-typescript": "^9.0.0",
     "mocha": "^5.2.0",
     "npm-run-all": "^4.1.2",
-    "nyc": "^13.1.0",
     "power-assert": "^1.4.4",
     "rimraf": "^2.6.2",
-    "source-map-support": "^0.5.9",
     "supervisor": "^0.12.0",
     "ts-node": "^7.0.1",
     "tslint": "^5.11.0",
-    "typedoc": "^0.11.1",
-    "typescript": "^2.9.2"
+    "typedoc": "^0.13.0",
+    "typescript": "^3.1.6"
   },
   "directories": {
     "test": "test"
@@ -76,31 +80,14 @@
     "compile:ts": "tsc --project .",
     "doc": "typedoc --mode modules --excludeExternals --ignoreCompilerErrors --exclude \"**/*.d.ts\" --out doc index.ts lib",
     "gql:gen": "atm-gql-gen",
-    "lint": "tslint --format verbose --project . --exclude \"node_modules/**\" --exclude \"**/*.d.ts\" \"**/*.ts\"",
+    "lint": "tslint --config tslint.json --format verbose --project .",
     "lint:fix": "npm run lint -- --fix",
-    "test": "nyc mocha --require espower-typescript/guess --require source-map-support/register \"test/**/*.test.ts\"",
+    "test": "mocha --require espower-typescript/guess --require source-map-support/register \"test/**/*.test.ts\"",
     "test:one": "mocha --require espower-typescript/guess \"test/**/${TEST:-*.test.ts}\"",
     "typedoc": "npm run doc"
   },
   "engines": {
     "node": ">=8.1.0",
     "npm": ">=5.0.0"
-  },
-  "nyc": {
-    "include": [
-      "lib/**/*.*"
-    ],
-    "extension": [
-      ".ts"
-    ],
-    "exclude": [
-      "lib/typings",
-      "**/*.d.ts"
-    ],
-    "reporter": [
-      "html",
-      "text-summary"
-    ],
-    "all": true
   }
 }
